backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
psutil>=5.9.0
PyJWT>=2.8.0
cryptography>=41.0.0
GPUtil>=1.4.0
requests>=2.31.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
peft>=0.4.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
# Apple Silicon support (optional)
mlx>=0.8.0
mlx-lm>=0.8.0

# Avatar training dependencies
diffusers>=0.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
Pillow>=9.0.0
opencv-python>=4.5.0
pyyaml>=6.0
jsonschema>=4.0.0
# xformers>=0.0.20  # CUDA only, install separately if needed
